concluded

+++ DS_09_measurement_converter.py
@@ -1,5 +1,5 @@
 print('\033[1;32m Enter a distance in meters and find your conversion')
-m = int(input('Distance in meters:'))
+m = float(input('Distance in meters:'))
 km = m / 1000
 print (f'\033[4;34m Kilometers: {km}km')
 dam = m / 10
